add_repo_to_env refactoring
REASON: We need to have that task compatible with future solution of
https://bugs.launchpad.net/mos/+bug/1650500 in nailgun.

FIX: Removed shell. Only py script is in charge. Ansible can't work
when script called script so we need only one.

+++ playbooks/files/add_repo_to_env.py
@@ -2,7 +2,16 @@
 
 import io
 import yaml
+import os
+import shutil
 import sys
+
+def get_from_env(name):
+    if name in os.environ:
+        return os.environ[name]
+    else:
+        sys.exit("Please specify " + name)
+
 
 def priority(value):
     try: 
@@ -10,16 +19,49 @@ def priority(value):
     except ValueError:
         return None
 
-with io.open(sys.argv[1], "r") as ifile:
-    data=yaml.load(ifile)
+
+def run_or_die(cmd):
+    if os.system(cmd)!=0:
+        sys.exit(cmd + "has been failed with")
+
+
+def repo_exists(repos,name):
+    for repo in repos:
+        if 'name' in repo:
+            if repo['name']==name:
+                return True
+    return False
+
+
+def main():
+    env_id=get_from_env("ENV_ID")
+    name=get_from_env("NAME")
+    filename="cluster_{}/attributes.yaml".format(env_id)
+
+    os.chdir("/tmp")
+    run_or_die("fuel env --env {} --attributes --download".format(env_id))
+    shutil.copyfile(filename, filename+".orig")
+
+    with io.open(filename, "r") as ifile:
+        data=yaml.load(ifile)
+
+    if repo_exists(data['editable']['repo_setup']['repos']['value'], name):
+        print "Repo with name " + name + " already exists"
+        sys.exit()
 
     data['editable']['repo_setup']['repos']['value'].append(
-            {'name':    sys.argv[2],
-            'priority': priority(sys.argv[3]),
-            'section':  sys.argv[4],
-            'suite':    sys.argv[5],
-            'type':     sys.argv[6],
-            'uri':      sys.argv[7],
+            {'name':    name,
+            'priority': priority(get_from_env("PRIORITY")),
+            'section':  get_from_env('SECTION'),
+            'suite':    get_from_env('SUITE'),
+            'type':     get_from_env('TYPE'),
+            'uri':      get_from_env('URL'),
         })
-    with io.open(sys.argv[1], "w") as ofile:
+
+    with io.open(filename, "w") as ofile:
         yaml.dump(data,ofile, default_flow_style=False)
+
+    run_or_die("fuel env --env {} --attributes --upload".format(env_id))
+
+if __name__ == '__main__':
+    main()
